+++ __init__.py
@@ -1,9 +1,15 @@
+# __init__.py
+
 import os
+
+# 导入所有节点类
 from .nodes.custom_save_image import CustomSaveImage
 from .nodes.color_adjustment import ColorAdjustment
 from .nodes.strong_prompt import StrongPrompt
 from .nodes.baidu_translater import BaiduTranslater
 from .nodes.display_any import DisplayAny  # 添加新的 DisplayAny 节点
+from .nodes.abc_math import AbcMath  # 确保路径正确
+from .nodes.image_size_extractor import ImageSizeExtractor  # 添加 ImageSizeExtractor 节点
 
 # 节点类映射
 NODE_CLASS_MAPPINGS = {
@@ -11,7 +17,9 @@ NODE_CLASS_MAPPINGS = {
     "Color_Adjustment": ColorAdjustment,
     "Strong_Prompt": StrongPrompt,
     "Baidu_Translater": BaiduTranslater, 
-    "Display_Any": DisplayAny,  # 添加新的 DisplayAny 节点映射
+    "Display_Any": DisplayAny,
+    "Abc_Math": AbcMath,
+    "Image_Size_Extractor": ImageSizeExtractor,  # 添加 ImageSizeExtractor 映射
 }
 
 # 节点显示名称映射
@@ -20,10 +28,10 @@ NODE_DISPLAY_NAME_MAPPINGS = {
     "Color_Adjustment": "Color Adjustment",
     "Strong_Prompt": "Strong Prompt",
     "Baidu_Translater": "Baidu Translater", 
-    "Display_Any": "Display Any",  # 添加新的 DisplayAny 显示名称映射
+    "Display_Any": "Display Any",
+    "Abc_Math": "abc Math",  # 显示名称映射
+    "Image_Size_Extractor": "Image Size Extractor",  # 添加 Image Size Extractor 显示名称映射
 }
 
 # 定义 Web 目录路径
 WEB_DIRECTORY = "web"
-
-# 如果有其他初始化代码或设置，可以继续添加在这里